requirements.txt
Typing
Pillow